Warning fixes and code cleanup
- Remove unused imports
- Remove unnecessary casts
- Remove unused variables
- Remove throws declarations for exceptions that aren't actually thrown
- Add @SafeVarargs where appropriate
- Make Enums more type-safe
- Avoid using both a lambda and a separate method
- Remove extraneous semicolons

+++ src/main/java/mcjty/theoneprobe/apiimpl/providers/DefaultProbeInfoEntityProvider.java
@@ -164,7 +164,7 @@ public class DefaultProbeInfoEntityProvider implements IProbeInfoEntityProvider 
         if (effect.getDuration() == 32767) {
             return "**:**";
         } else {
-            int i = MathHelper.floor((float)effect.getDuration() * factor);
+            int i = MathHelper.floor(effect.getDuration() * factor);
             return ticksToElapsedTime(i);
         }
     }
